backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydantic==2.7.4
pydantic[email]==2.7.4
scikit-learn==1.5.0
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
spacy==3.8.2
pypdf==4.2.0
reportlab==4.2.0
requests==2.32.3
python-dotenv==1.0.1
